stop indefinite hangs in non-cors environments
There might be a better way to handle this, but it was an interesting discovery to find that without the CORS block error message the request would hang indefinitely on invalid URLs. Not all URLs though, I’ve implemented the subset above (that I found) to avoid breaking any weird backward compatibility that may arise.